# Greatly speed up adapter counter by using out of order execution