--- a/src/sequali/_qcmodule.c
+++ b/src/sequali/_qcmodule.c
@@ -1268,10 +1268,8 @@ typedef struct AdapterCounterStruct {
     PyObject *adapters;
     size_t number_of_matchers;
     MachineWordPatternMatcher *matchers;
-#ifdef __SSE2__
     size_t number_of_sse2_matchers;
     MachineWordPatternMatcherSSE2 *sse2_matchers;
-#endif
 } AdapterCounter;
 
 static void AdapterCounter_dealloc(AdapterCounter *self) {
@@ -1440,10 +1438,8 @@ AdapterCounter__new__(PyTypeObject *type, PyObject *args, PyObject *kwargs)
     self->number_of_adapters = number_of_adapters;
     self->number_of_matchers = 0;
     self->number_of_sequences = 0;
-    #ifdef __SSE2__ 
     self->number_of_sse2_matchers = 0;
     self->sse2_matchers = NULL;
-    #endif
     size_t adapter_index = 0;
     size_t matcher_index = 0;
     PyObject *adapter;