### Checklist
- [ ] Pull request details were added to CHANGELOG.rst
- [ ] Documentation was updated (if needed)

The shift-AND algorithm creates a dependency chain, each computation relies on a result by a former computation. As a result  this is very hard to optimize.

By running two Shift-AND matchers in the same loop, they can be evaluated independently. Modern CPU's can execute these statements out of order leading to simultaneous evaluation. 

The current adapter matching code matches 15 12bp pieces to each sequence when the file is identified as nanopore. That requires three 64-bit integers to store all the pieces. Two of those are combined in a SSE2 vector. Evaluating all the sequencing takes about 650ms on my machine using a 64-bit vector and 850 ms using a SSE2 vector. The total cost for running the routine was thus 850 + 650 ms: roughly 1500 ms.  However thanks to the SSE2 vector and the 64-bit int running simultaneously the cost is now 850 ms, meaning the cost of running the 64-bit integer is entirely negated.

This makes sense from a hardware standpoint as a cpu has 16 general purpose 64 bit registers and 16 vector registers (for x86-64) and can support out of order execution. Nevertheless it is a very impressive result. 

Application performance (all modules!) before:
```
Benchmark 1: sequali ~/test/nanopore_100000reads.fastq
  Time (mean ± σ):      5.633 s ±  0.060 s    [User: 5.339 s, System: 0.293 s]
  Range (min … max):    5.538 s …  5.685 s    5 runs
```

Application performance after:
```
Benchmark 1: sequali ~/test/nanopore_100000reads.fastq
  Time (mean ± σ):      5.086 s ±  0.040 s    [User: 4.805 s, System: 0.279 s]
  Range (min … max):    5.045 s …  5.134 s    5 runs
```

On larger files where the report generation takes less of the total time the improvement is even better.